added message builder class
