+++ MessageQueueManager/Interfaces/IMessageQueueManager.cs
@@ -2,7 +2,6 @@
 {
     public interface IMessageQueueManager
     {
-        bool CreateMessageQueue(string queueName);
         bool SendMessage(string message);
         string ReadMessage();
     }
